MLS test additions and cleanups (#2285)
* Add assertion to MLS commit test

* Move creation tests to MLS module

* Split messaging test setup to its own module

* Remove unused function

* Clean up application messaging tests

* Add CHANGELOG entry
