+++ libs/wire-api/src/Wire/API/MLS/Message.hs
@@ -26,7 +26,6 @@ module Wire.API.MLS.Message
     SWireFormatTag (..),
     SomeMessage (..),
     ContentType (..),
-    mpTag,
     MessagePayload (..),
     MessagePayloadTBS (..),
     Sender (..),
@@ -161,12 +160,6 @@ data ContentType
 
 instance ParseMLS ContentType where
   parseMLS = parseMLSEnum @Word8 "content type"
-
--- | Identify the content type of a message payload to be signed.
-mpTag :: MessagePayloadTBS -> ContentType
-mpTag (ApplicationMessage _) = ApplicationMessageTag
-mpTag (ProposalMessage _) = ProposalMessageTag
-mpTag (CommitMessage _) = CommitMessageTag
 
 instance ParseMLS MessagePayloadTBS where
   parseMLS =
