+++ services/galley/test/integration/API.hs
@@ -122,8 +122,6 @@ tests s =
         [ test s "status" status,
           test s "metrics" metrics,
           test s "create Proteus conversation" postProteusConvOk,
-          test s "fail to create MLS conversation" postMLSConvFail,
-          test s "create MLS conversation" postMLSConvOk,
           test s "create conversation with remote users" postConvWithRemoteUsersOk,
           test s "get empty conversations" getConvsOk,
           test s "get conversations by ids" getConvsOk2,
@@ -286,34 +284,6 @@ postProteusConvOk = do
       case evtData e of
         EdConversation c' -> assertConvEquals cnv c'
         _ -> assertFailure "Unexpected event data"
-
-postMLSConvFail :: TestM ()
-postMLSConvFail = do
-  qalice <- randomQualifiedUser
-  let alice = qUnqualified qalice
-  bob <- randomUser
-  connectUsers alice (list1 bob [])
-  postConvQualified alice defNewMLSConv {newConvQualifiedUsers = [Qualified bob (qDomain qalice)]}
-    !!! do
-      const 400 === statusCode
-      const (Just "non-empty-member-list") === fmap label . responseJsonError
-
-postMLSConvOk :: TestM ()
-postMLSConvOk = do
-  c <- view tsCannon
-  qalice <- randomQualifiedUser
-  bob <- randomUser
-  let alice = qUnqualified qalice
-  let nameMaxSize = T.replicate 256 "a"
-  connectUsers alice (list1 bob [])
-  WS.bracketR2 c alice bob $ \(wsA, wsB) -> do
-    rsp <- postConvQualified alice defNewMLSConv {newConvName = checked nameMaxSize}
-    pure rsp !!! do
-      const 201 === statusCode
-      const Nothing === fmap label . responseJsonError
-    cid <- assertConv rsp RegularConv alice qalice [] (Just nameMaxSize) Nothing
-    WS.assertNoEvent (2 # Second) [wsB]
-    checkConvCreateEvent cid wsA
 
 postConvWithRemoteUsersOk :: TestM ()
 postConvWithRemoteUsersOk = do
